api/requirements.txt
asgiref==3.5.2
astroid==2.12.10
black==22.8.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dill==0.3.5.1
Django==4.0.5
django-mysql==4.7.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysql-connector-repackaged==0.3.1
mysqlclient==2.1.1
pathspec==0.10.1
platformdirs==2.5.2
protobuf==4.21.6
pycparser==2.21
pylint==2.15.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
